Peak cleaning algorithm, constrain parameters
Added a new function to filter peaks - remove small ones that are likely
to be in the debris field, pick the smaller of a pair of peaks that are
likely to be a G1/G2 pair, and drop short peaks that are part of the
noise.

Also switched to the port algorithm, which allows for constraining
parameters. With this option, parameters can be restricted to be >= 0.
However, I have now discovered a problem - when the G2 peak is off
scale (ie., doesn't appear in the data), the parameters controlling that
peak have no data to be fit too, so can produce wildly unrealistic
values. These values can muck up estimation of confidence intervals, so
I should probably rebuild the models so that they don't include the G2
peaks at all.